--- Chat.java
@@ -1,5 +1,0 @@
-import java.util.ArrayList;
-public class Chat {
-    ArrayList<User> users = new ArrayList<>();
-    Integer noOfMembers = 0;
- }
