--- User.java
@@ -1,7 +1,0 @@
-import java.util.ArrayList;
-
-public class User {
-    String password;
-    String username;
-    String bio;
-}
