+++ src/entity/Chat.java
@@ -1,0 +1,9 @@
+package entity;
+
+import java.util.ArrayList;
+
+public class Chat {
+    ArrayList<User> users = new ArrayList<>();
+    Integer noOfMembers;
+    ArrayList<Message> allmessages = new ArrayList<>();
+}
