+++ src/entity/Message.java
@@ -1,0 +1,7 @@
+package entity;
+
+public class Message {
+    String Photo;
+    User sender;
+    User receiver;
+}
