+++ src/entity/Requests.java
@@ -1,0 +1,6 @@
+package entity;
+
+public class Requests {
+    User From;
+    User To;
+}
